update button position
